% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_stratified.R
\name{eval_stratified}
\alias{eval_stratified}
\title{Evalueer samennemend de resultaten van 1 of meer steekproeven op uitgaand geld}
\usage{
eval_stratified(
  steekproeven,
  zekerheid = 0.95,
  MC = 1e+07,
  start = 1,
  vergelijk = TRUE
)
}
\arguments{
\item{steekproeven}{Een tibble.
Elke regel van de tibble beschrijft 1 steekproef, dus 1 van de genomen
steekproeven.
De tibble heeft als kolommen: naam w n k car cir dir.
Dit zijn repectievelijk
\code{naam}, een aanduiding van de steekproef,
\code{w}, de omvang in geld van de massa waaruit getrokken is,
\code{n}, het aantal getrokken posten,
\code{k}, de som van de foutfracties van de posten,
\code{ihr}, inherent risico, te weten H, M of L,
\code{ibr}, intern beheersingsrisico, te weten H, M of L en
\code{car}, cijferanalyserisico, te weten H, M of L.}

\item{zekerheid}{Het zekerheidsniveau waarop we de maximale foutfractie berekenen.}

\item{MC}{Het aantal Monte Carlo iteraties dat gebruikt wordt.
Monte Carlo berekeningen baseren zich op toevalsgetallen.}

\item{start}{Startwaarde voor de toevalsgenerator.}

\item{vergelijk}{TRUE of FALSE, als TRUE dan worden wat vergelijkende berekeningen uitgevoerd
en de resultaten daarvan toegevoegd aan de uitkomst van de functie.}
}
\value{
Een lijst, bestaande uit
\itemize{
\item \code{mw_fout_convolutie}, de meest waarschijnlijke fout als fractie van de totale massa in geld
\item \code{max_fout_convolutie}, de maximale fout als fractie van de totale massa in geld
\item en zo ook voor modus, mediaan en gemiddelde
\item als vergelijk == TRUE een lijst
\code{vergelijk_met}, met daarin vergelijkende cijfers voor "los" en voor "als1".
Hierbij staat "los" voor de samengenomen, gewogen, cijfers der afzonderlijke steekproeven,
en "als1" voor als alle steken en de resultaten daarvan worden beschouwd
als voor 1 steekproef getrokken uit 1 massa.
\item \code{steekproeven}, de invoer tibble, verrijkt per steekproef: met
\itemize{
\item \code{extra_foutloze_posten}, het aantal foutloze posten dat equivalent is aan ihr+ibr+car volgens
het ARM model en de interpretatie daarvan door de ADR,
\item \code{toch_fouten}, TRUE of FALSE, geeft aan of tenminste 1 van ihr, ibr of car niet op hoog staat
en er wel fouten zijn gevonden
\item \code{mw_fout}, als vergelijk TRUE,
de meest waarschijnlijke fout voor die steekproef als
fractie van de massa in geld van de steekproef,
\item \code{max_fout}, als vergelijk TRUE,
de maximale fout als fractie van de massa in geld van de steekproef,
}
\item \code{invoer}, een lijst bestaande uit de invoerparameters
}
}
\description{
Het samennemen van de resultaten gebeurt door convolutie van
de foutkanskrommes van de afzonderlijke steekproeven tot
1 foutkanskromme.

We berekenen de maximale en meest waarschijnlijke fout als fractie
en in geld van de samengenomen geldbedragen.

De maximale fout wordt bepaald aan de hand van de resulterende kanskromme, op basis
van de gewenste zekerheid. Visueel is de maximale fout, pm, te bepalen in een
tweedimensionaal, haaks, assenstelsel.
De horizontale as, de p-as, loopt van 0 tot 1.
De waarden langs die as geven de mogeljke foutfracties weer,
lopend van 0 (geen fouten) tot 1 (alles fout).
De verticale as, de c-as, loopt van 0 to oneindig.
Deze as geeft de kanswaarden van de foutfracties aan.
In dit assenstelsel kunnen we de kanskromme afbeelden.
Het oppervlak onder de kanskromme is 1.
Hierbij praten we over het oppervlak begrenst door de p-as, aan de onderkant,
en de verticale lijnen p = 0, en p = 1.
pm is het punt op de p-as waarbij de verticale lijn p = pm,
het oppervlak onder de kanskromme begrenst zodat links van deze lijn
het oppervlak gelijk is aan de zekerheid, bijvoorbeeld 0,95.

De meest waarschijnlijke fout is het gewogen gemiddelde van de p-waarden
van alle punten op de kanskromme.
De weging is op basis van de bijbehorende kans (c) waarden.
}
\details{
We gaan hierbij uit van de som van de foutfracties, dus we kijken niet naar
de foutfracties per post.

Aggregatie is dus puur op statistische
gronden: namelijk risico's op fouten boven de meest waarschijnlijke fout
en op onder de meest waarschijnlijke fout vlakken elkaar enigszins uit
genomen over de meerdere steekproeven.
Dus, bij het aggregeren van de resultaten van de verschillende steekproeven
wordt geen enkele aanname gedaan over gelijkenis tussen
de eigenschappen van de afzonderlijke administraties waaruit is
getrokken.

Dit package kan ook steekproeven combineren over massa's waarvoor
een verschillende risicoinschatting geldt.

Bijvoorbeeld:
Steekproef1 is gebaseerd op een zekerheid van 95\% omdat
ihr, ibr en car alledrie op hoog (H) staan.
De materialiteit is 2\%.
Het betreft 100 miljoen euro.
Voor steekproef1 trekken we 148 posten, waarbij 1 fout blijkt.
Steekproef2 is gebaseerd op een zekerheid van 64\% omdat
ihr en ibr allebei op laag staan en alleen car op hoog.
Het betreft ook 100 miljoen euro en een materialiteit van 2\%.
Voor steekproef2 trekken we 50 posten waarvan er 0 fout blijken.

Bij verschil in risicoinschatting van de massa's waarover
wordt gestoken worden de lagere risicoinschattingen vertaald naar
extra getrokken foutloze posten.
In ons voorbeeld bepalen we voor steekproef2 het aantal foutloze
posten nodig om een positieve uitspraak te doen bij 64\% en bij 95\%.
Dit zijn respectievelijk 50 en 148.
Het verschil is 148-50 = 98 posten.
Daarna berekenen we totale maximale fout op basis van
een zekerheid van 95\%,
en steekproef1, 148 posten waarvan 1 fout,
en steekproef2, met 50 + 98 posten, waarvan 0 fout.
De maximale fout is dan 1,83\% ofwel ongeveer3.660.000.
De meest waarschijnlijke fout 0,52\% ofwel ongeveer 1.160.000 euro.

Het is de verantwoordelijkheid van de auditor hoe om te gaan
met een steekproef waarbij de risicoinschatting niet op H staat
en er toch fouten worden gevonden. Dit probleem staat los
van hoe de uitkomsten van meerdere steekproeven samen te nemen.

Als de parameter vergelijk TRUE is doen we, ter vergelijking, ook de evaluatie:
\itemize{
\item voor elke steekproef los
\item voor alle steekproeven samen, waarbij ze beschouwd worden als te zijn getrokken op 1 massa.
}
}
\examples{
steekproeven <- tribble(~naam, ~w, ~n, ~k) # Creeer lege invoertibble.

naam <- "Steekproef1" # De naam van de 1e steekproef.
w <- 35060542 # Het gewicht van de steekproef, als geldomvang van de massa waarover wordt gestoken.
n <- 31 # Het aantal getrokken en geevalueerde posten.
k <- 0 # De som van de foutfracties.
steekproeven <- add_row(steekproeven, naam = naam, w = w, n = n, k = k)

naam <- "Steekproef2" # De naam van de 2e steekproef.
w <- 3044699  # Het gewicht van de steekproef, als geldomvang van de massa waarover wordt gestoken.
n <- 8 # Het aantal getrokken en geevalueerde posten.
k <- 0 # De som van de foutfracties.
steekproeven <- add_row(steekproeven, naam = naam, w = w, n = n, k = k)

# Kortheidshalve hadden we steekproeven ook kunnen invoeren als:
steekproeven <- tribble(
~naam, ~w, ~n, ~k, ~ihr, ~ibr, ~car,
"Steekproef1", 35060542, 31, 0, 'H', 'H', 'H',
"Steekproef2", 3044699, 8, 0, 'H', 'H', 'H')

# Evalueer steekproef1 en steekproef2 samen.
eval_stratified(steekproeven = steekproeven)
}
