% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalstratified.R
\name{evalstratified}
\alias{evalstratified}
\title{Evalueer samennemend de resultaten van 1 of meer steekproeven op uitgaand geld}
\usage{
evalstratified(steken, zekerheid = 0.95, MC = 1e+07, start = 1)
}
\arguments{
\item{steken}{Een tibble.
Elke regel van de tibble beschrijft 1 steekproef.
De tibble heeft als kolommen naam w n k car cir dir.
Dit zijn repectievelijk
\code{naam}, een aanduiding, van de steekproef,
\code{w}, de omvang in geld van de massa waaruit getrokken is,
\code{n}, het aantal getrokken posten,
\code{k}, de som van de foutfracties van de posten,
\code{ihr}, ihr risico, te weten H, M of L,
\code{ibr}, ibr risico, te weten H, M of L en
\code{car}, car risico, te weten H, M of L.}

\item{zekerheid}{Het zekerheidsniveau waarop we de maximale foutfractie berekenen.}

\item{MC}{Het aantal Monte Carlo berekeningen per steekproef, berekeningen gebaseerd op toevalsgetallen, dat gebruikt wordt.}

\item{start}{Startwaarde voor de toevalsgenerator.}
}
\value{
Een lijst, bestaande uit
\itemize{
\item \code{mw_fout_samen_onafh}, de meest waarschijnlijke fout als fractie van de totale massa in geld
\item \code{mw_fout_samen_onafh_geld}, de meest waarschijnlijke fout in geld
\item \code{max_fout_samen_onafh}, de maximale fout als fractie van de totale massa in geld
\item \code{max_fout_samen_onafh_geld}, de maximale fout in geld
\item een tibble, \code{steken} waarin per steekproef:
\itemize{
\item \code{naam}, een aanduiding, van de steekproef,
\item \code{w}, de omvang in geld van de massa waaruit getrokken is,
\item \code{n}, het aantal getrokken posten,
\item \code{k}, de som van de foutfracties van de posten,
\item \code{ihr}, het ihr risico,
\item \code{ibr}, het ibr risico,
\item \code{car}, het car risico,
\item \code{extra_foutloze_posten}, het aantal foutloze posten dat equivalent is aan ihr+ibr+car volgens
het ARM model en de interpretatie daarvan door de ADR,
\item \code{mw_fout}, de meest waarschijnlijke fout als fractie van de massa in geld van de steekproef,
\item \code{mw_fout_geld}, de meest waarschijnlijke fout in geld,
\item \code{max_fout}, de maximale fout als fractie van de massa in geld van de steekproef,
\item \code{max_fout_geld}, de maximale fout in geld.
}
\item \code{zekerheid}
\item \code{MC}
\item \code{start}
\item \code{mw_fout_samen_afh}, de meest waarschijnlijke fout als percentage, er
vanuit gaande dat de steekproeven allemaal samen uit dezelfde massa zouden zijn getrokken
\item \code{mw_fout_samen_afh_geld}, de meest waarschijnlijke fout in geld daarbij
\item \code{max_fout_samen_afh}, de maximale fout als percentage, er
vanuit gaande dat de steekproeven allemaal samen uit dezelfde massa zouden zijn getrokken
\item \code{max_fout_samen_afh_geld}, de maximale fout in geld daarbij
}
}
\description{
Evalueer samennemend als onafhankelijke steekproeven 1 of meer steekproeven op
uitgaand geld. Doe dit ook door ze te beschouwen als getrokken op dezelfde massa,
dus als afhankelijk.
Evalueer ook de steekproeven afzonderlijk.
}
\details{
We gaan hierbij uit van de som van de foutfracties, dus we kijken niet naar
de foutfracties per post.
Belangrijkste dat we berekenen is de maximale foutfractie.
}
\examples{
steken <- tribble(~naam, ~w, ~n, ~k) # Creeer lege invoertibble.

naam <- "Steekproef1" # De naam van de 1e steekproef.
w <- 35060542 # Het gewicht van de steekproef, als geldomvang van de massa waarover wordt gestoken.
n <- 31 # Het aantal getrokken en geevalueerde posten.
k <- 0 # De som van de foutfracties.
steken <- add_row(steken, naam = naam, w = w, n = n, k = k)

naam <- "Steekproef2" # De naam van de 2e steekproef.
w <- 3044699  # Het gewicht van de steekproef, als geldomvang van de massa waarover wordt gestoken.
n <- 8 # Het aantal getrokken en geevalueerde posten.
k <- 0 # De som van de foutfracties.
steken <- add_row(steken, naam = naam, w = w, n = n, k = k)

# Kortheidshalve hadden we steken ook kunnen invoeren als:
steken <- tribble(
~naam, ~w, ~n, ~k, ~ihr, ~ibr, ~car,
"Steekproef1", 35060542, 31, 0, 'H', 'H', 'H',
"Steekproef2", 3044699, 8, 0, 'H', 'H', 'H')

# Evalueer steekproef1 en steekproef2 samen.
evalstratified(steken = steken)
}
