% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foutloze_posten_equivalent.R
\name{foutloze_posten_equivalent}
\alias{foutloze_posten_equivalent}
\title{Geef equivalent in getrokken foutloze posten voor verlaagde risicoschattingen.}
\usage{
foutloze_posten_equivalent(
  ihr = "H",
  ibr = "H",
  car = "H",
  materialiteit = 0.01
)
}
\arguments{
\item{ihr}{inherente risico, te weten H, M of L}

\item{ibr}{internebeheersingsrisico, te weten H, M of L}

\item{car}{cijferanalyserisico, te weten H, M of L}

\item{materialiteit}{de maximale foutfractie in de te onderzoeken geldmassa}
}
\value{
Het equivalent in getrokken foutloze posten. Een integer >= 0.
}
\description{
Geef equivalent in getrokken foutloze posten voor verlaagde risicoschattingen.
}
\examples{
fpe <- foutloze_posten_equivalent()

}
